backend/scripts/requirements-training.txt
# =============================================================
# Training dependencies for SmolVLM fine-tuning on RTX 4060 Mobile
# Install:  pip install -r scripts/requirements-training.txt
# =============================================================

# --- PyTorch (CUDA 12.x) ---
--index-url https://download.pytorch.org/whl/cu124
torch>=2.4.0
torchvision>=0.19.0

# --- HuggingFace ecosystem ---
transformers>=4.49.0
datasets>=3.0.0
accelerate>=1.2.0
peft>=0.14.0
huggingface-hub>=0.33.0

# --- Quantisation (optional – QLoRA) ---
bitsandbytes>=0.45.0

# --- Image handling ---
Pillow>=10.0.0

# --- Metrics & evaluation ---
evaluate>=0.4.0
scikit-learn>=1.4.0

# --- Logging ---
tensorboard>=2.18.0

# --- Progress bars ---
tqdm>=4.66.0
